# CRM/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CombatDen API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/classes/checkin": {
      "post": {
        "tags": [
          "classes"
        ],
        "summary": "Check a member into a class instance",
        "description": "Selects the best eligible membership plan with remaining capacity (trial -> one_time -> recurring), logs the attendance against it, bumps ``last_class``, and auto-ends trial / punch-card memberships once depleted. Returns the membership breakdown. If no plan covers the class with capacity, the check-in is rejected: ``log_id`` is null and no attendance is written. Idempotent \u2014 a repeat for the same (member_id, class_history_id) returns the existing log_id with ``already_checked_in = True`` and consumes no capacity.",
        "operationId": "checkin_api_v1_classes_checkin_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Check-in result returned (recorded or rejected)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckinResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized for this member"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/classes/streak": {
      "get": {
        "tags": [
          "classes"
        ],
        "summary": "Get a member's class attendance streak",
        "description": "Weeks of consecutive class attendance.",
        "operationId": "get_streak_api_v1_classes_streak_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "member_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Member Id"
            }
          },
          {
            "name": "gym_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Gym Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Streak retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreakResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized for this member"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckinMembershipBreakdown": {
        "properties": {
          "plan_id": {
            "type": "string",
            "format": "uuid",
            "title": "Plan Id"
          },
          "plan_type": {
            "$ref": "#/components/schemas/PlanType"
          },
          "class_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Class Count"
          },
          "classes_used": {
            "type": "integer",
            "title": "Classes Used"
          },
          "classes_remaining": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classes Remaining"
          },
          "is_eligible": {
            "type": "boolean",
            "title": "Is Eligible"
          },
          "renew_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Renew Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          }
        },
        "type": "object",
        "required": [
          "plan_id",
          "plan_type",
          "class_count",
          "classes_used",
          "classes_remaining",
          "is_eligible",
          "renew_date",
          "end_date"
        ],
        "title": "CheckinMembershipBreakdown",
        "description": "Usage breakdown for one of the member's active memberships.\n\nAttributes:\n    plan_id: The membership plan.\n    plan_type: Trial, one_time, or recurring.\n    class_count: Max classes allowed (None = unlimited).\n    classes_used: Classes used this cycle (post-checkin for the\n        charged plan).\n    classes_remaining: Classes left (None = unlimited).\n    is_eligible: Whether this plan covers the checked-in class.\n    renew_date: Next renewal / due date (recurring plans; None\n        otherwise).\n    end_date: Membership expiry date (trial / one_time plans; None\n        otherwise)."
      },
      "CheckinRequest": {
        "properties": {
          "member_id": {
            "type": "string",
            "format": "uuid",
            "title": "Member Id"
          },
          "gym_id": {
            "type": "string",
            "format": "uuid",
            "title": "Gym Id"
          },
          "class_history_id": {
            "type": "string",
            "format": "uuid",
            "title": "Class History Id"
          }
        },
        "type": "object",
        "required": [
          "member_id",
          "gym_id",
          "class_history_id"
        ],
        "title": "CheckinRequest",
        "description": "Body for POST /api/v1/classes/checkin."
      },
      "CheckinResponse": {
        "properties": {
          "log_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Log Id"
          },
          "member_id": {
            "type": "string",
            "format": "uuid",
            "title": "Member Id"
          },
          "class_history_id": {
            "type": "string",
            "format": "uuid",
            "title": "Class History Id"
          },
          "already_checked_in": {
            "type": "boolean",
            "title": "Already Checked In"
          },
          "chosen_plan_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chosen Plan Id"
          },
          "chosen_item_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chosen Item Id"
          },
          "memberships": {
            "items": {
              "$ref": "#/components/schemas/CheckinMembershipBreakdown"
            },
            "type": "array",
            "title": "Memberships",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "log_id",
          "member_id",
          "class_history_id",
          "already_checked_in"
        ],
        "title": "CheckinResponse",
        "description": "Response for POST /api/v1/classes/checkin.\n\nThe check-in is gated: a covering active membership with remaining\ncapacity must exist or the check-in is rejected (``log_id`` /\n``chosen_plan_id`` / ``chosen_item_id`` are ``None`` and no\nattendance row is written). The ``memberships`` breakdown explains\nthe decision either way.\n\nAttributes:\n    log_id: The attendance row. None when the check-in was rejected.\n    member_id: The member who checked in.\n    class_history_id: The class instance attended.\n    already_checked_in: True when an attendance row already existed\n        for this (member, class instance) \u2014 the check-in was\n        idempotent and no capacity was consumed.\n    chosen_plan_id: The plan charged. None when rejected.\n    chosen_item_id: The membership row charged. None when rejected.\n    memberships: Breakdown of the member's active memberships."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlanType": {
        "type": "string",
        "enum": [
          "trial",
          "one_time",
          "recurring"
        ],
        "title": "PlanType",
        "description": "Membership plan type, matching the DB CHECK constraint."
      },
...
      "StreakResponse": {
        "properties": {
          "member_id": {
            "type": "string",
            "format": "uuid",
            "title": "Member Id"
          },
          "class_streak_weeks": {
            "type": "integer",
            "title": "Class Streak Weeks"
          }
        },
        "type": "object",
        "required": [
          "member_id",
          "class_streak_weeks"
        ],
        "title": "StreakResponse",
        "description": "Response for GET /api/v1/classes/streak."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}